Refactor MultiStaticSelectElement
